remover paciente e alterarPaciente

+++ src/main/java/com/hackaton/service/PacienteService.java
@@ -4,6 +4,8 @@ import com.hackaton.dto.PacienteDTO;
 import com.hackaton.entity.Paciente;
 import com.hackaton.exception.ConflitoException;
 import com.hackaton.repository.PacienteRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -12,6 +14,7 @@ import org.springframework.stereotype.Service;
 public class PacienteService {
 
 
+    private static final Logger log = LoggerFactory.getLogger(PacienteService.class);
     private final PacienteRepository pacienteRepository;
 
     public PacienteService(PacienteRepository pacienteRepository) {
@@ -51,6 +54,31 @@ public class PacienteService {
     }
 
     public void removerPaciente(Long id) {
-        pacienteRepository.deleteById(id);
+
+        if (!pacienteRepository.existsById(id)) {
+            throw new IllegalArgumentException("Paciente não encontrado.");
+        }
+        try {
+            pacienteRepository.deleteById(id);
+            log.info("Paciente com id {} removido com sucesso",id);
+        }catch (Exception e){
+            log.error("Erro ao remover paciente com id {}: {}",id, e.getMessage());
+            throw new RuntimeException("Erro ao remover paciente",e);
+        }
+
+    }
+
+    public Paciente alteraPaciente(PacienteDTO pacienteDTO) {
+
+        Paciente pacienteExistente = pacienteRepository.findByCpf(pacienteDTO.cpf());
+        if (pacienteExistente == null) {
+            throw new IllegalArgumentException("Paciente não encontrado com o CPF informado.");
+        }
+
+        pacienteExistente.setNome(pacienteDTO.nome());
+        pacienteExistente.setCpf(pacienteDTO.cpf());
+        pacienteExistente.setEmail(pacienteDTO.email());
+
+        return pacienteRepository.save(pacienteExistente);
     }
 }
